+++ R/utils.R
@@ -3,12 +3,12 @@
 #' @importFrom tools R_user_dir
 #'
 getTempCacheDir <- function(create=TRUE) {
-  tempDir <- R_user_dir("PAMscapes", which = "cache")
-  if(create &&
-     !dir.exists(tempDir)) {
-    dir.create(tempDir, recursive=TRUE)
-  }
-  tempDir
+    tempDir <- R_user_dir("PAMscapes", which = "cache")
+    if(create &&
+       !dir.exists(tempDir)) {
+        dir.create(tempDir, recursive=TRUE)
+    }
+    tempDir
 }
 
 fileNameManager <- function(fileName=NULL, suffix=NULL) {
@@ -79,9 +79,9 @@ parseToUTC <- function(x,
     with_tz(origTz, tzone='UTC')
 }
 
-#' @importFrom tidyr pivot_longer
+#' @importFrom tidyr pivot_longer_spec pivot_longer
 #'
-toLong <- function(x) {
+toLong <- function(x, spec=NULL) {
     if(isLong(x)) {
         return(x)
     }
@@ -93,15 +93,13 @@ toLong <- function(x) {
     }
     if(type != 'BB') {
         freqCols <- gsub('[A-z]+_', '', colnames(x)[whichFreq])
-        # freqVals <- as.numeric(freqCols)
-        # names(freqVals) <- freqCols
         colnames(x)[whichFreq] <- freqCols
-        # colnames(x)[whichFreq] <- gsub('[A-z]+_', '', colnames(x)[whichFreq])
-        x <- pivot_longer(x, cols=all_of(whichFreq), names_to='frequency', values_to='value',
-                          names_transform = as.numeric)
-        # x$frequency <- as.numeric(x$frequency)
-        # x$frequency <- freqVals[x$frequency]
-        # x$type <- gsub('_[0-9-]+', '', x$type)
+        if(is.null(spec)) {
+            spec <- data.frame(.name=freqCols, .value='value', frequency=as.numeric(freqCols))
+        }
+        # x <- pivot_longer(x, cols=all_of(whichFreq), names_to='frequency', values_to='value',
+        #                   names_transform = as.numeric)
+        x <- pivot_longer_spec(x, spec=spec)
         x$type <- type
     }
     if(type == 'BB') {
@@ -127,14 +125,20 @@ toWide <- function(x) {
 }
 
 isLong <- function(x) {
-    all(c('UTC', 'type', 'value', 'frequency') %in% colnames(x))
+    if(is.data.frame(x)) {
+        x <- colnames(x)
+    }
+    all(c('UTC', 'type', 'value', 'frequency') %in% x)
 }
 
 isWide <- function(x) {
+    if(is.data.frame(x)) {
+        x <- colnames(x)
+    }
     freqCols <- whichFreqCols(x)
     # all(grepl('^[A-z]+_[0-9\\.\\-]+$', freqCols))
     if(length(freqCols) == 1 &&
-       grepl('BB', colnames(x)[freqCols])) {
+       grepl('BB', x[freqCols])) {
         return(TRUE)
     }
     length(freqCols) > 1
@@ -147,19 +151,30 @@ nowUTC <- function() {
 }
 
 
-#' @importFrom lubridate period
+#' @importFrom lubridate period period_to_seconds seconds_to_period is.period
 #'
 unitToPeriod <- function(x) {
+    if(is.period(x)) {
+        return(x)
+    }
     x <- gsub('([0-9]*)(.*)', '\\1_\\2', x)
     x <- strsplit(x, '_')[[1]]
     if(x[1] == '') {
         x[1] <- '1'
     }
-    period(as.numeric(x[1]), units=x[2])
+    # doing this to convert to roundest unit
+    # e.g. 720 seconds -> 12 minutes
+    seconds_to_period(
+        period_to_seconds(
+            period(as.numeric(x[1]), units=x[2])
+        )
+    )
 }
 
 whichFreqCols <- function(x) {
-    x <- colnames(x)
+    if(is.data.frame(x)) {
+        x <- colnames(x)
+    }
     isFreq <- grepl('^[A-z]+_[0-9\\.\\-]+$', x)
     colSplit <- split(x[isFreq], '_')
     type <- colSplit[[1]][1]
@@ -185,13 +200,13 @@ myLog10Scale <- function(g, range, dim=c('x', 'y')) {
     dim <- match.arg(dim)
     if(dim == 'x') {
         g +
-        scale_x_continuous(trans='log10',
-                           breaks = major,
-                           minor_breaks = minor,
-                           labels = scientific_10,
-                           limits = range,
-                           expand = c(0, 0)) +
-        annotation_logticks(sides='b')
+            scale_x_continuous(trans='log10',
+                               breaks = major,
+                               minor_breaks = minor,
+                               labels = scientific_10,
+                               limits = range,
+                               expand = c(0, 0)) +
+            annotation_logticks(sides='b')
     } else {
         g +
             scale_y_continuous(trans='log10',
@@ -262,4 +277,69 @@ typeToUnits <- function(type) {
            'HMD' = 'dB re: 1uPa^2/Hz',
            NULL
     )
+}
+
+colsToFreqs <- function(x) {
+    if(is.data.frame(x)) {
+        whichFreq <- whichFreqCols(colnames(x))
+        x <- colnames(x)[whichFreq]
+    }
+    as.numeric(gsub('[A-z]+_', '', x))
+}
+
+checkFreqType <- function(freq) {
+    if(is.character(freq)) {
+        nc <- nc_open(freq)
+        on.exit(nc_close(nc))
+        freq <- nc$dim$frequency$vals
+        if(is.null(freq)) {
+            warning('No frequency dimension')
+            return(NULL)
+        }
+    }
+    regDiff <- round(diff(freq), 1)
+    isOne <- regDiff == 1
+    if(all(isOne)) {
+        return('PSD')
+    }
+    # possible weirdness of others having 0, 1, 2? but not a lot
+    if(sum(isOne) > 2) {
+        return('HMD')
+    }
+
+    multDiff <- round(freq[2:length(freq)] / freq[1:(length(freq)-1)], 1)
+    if(all(multDiff == 2)) {
+        return('OL')
+    }
+    thirds <- seq(from=1, to=length(freq), by=3)
+    freq <- freq[thirds]
+    multDiff <- round(freq[2:length(freq)] / freq[1:(length(freq)-1)], 1)
+    if(all(multDiff == 2)) {
+        return('TOL')
+    }
+    warning('Could not parse frequency type')
+    'FREQ'
+}
+
+calcSliceLength <- function(dates, maxSlice) {
+    secDiff <- as.numeric(difftime(max(dates), min(dates), units='secs')) / maxSlice
+    # second, minute, hour, day are valid options
+    if(secDiff > 86400) {
+        value <- secDiff / 86400
+        unit <- 'day'
+    } else if(secDiff > 3600) {
+        value <- secDiff / 3600
+        unit <- 'hour'
+    } else if(secDiff > 60) {
+        value <- secDiff / 60
+        unit <- 'minute'
+    } else if(secDiff <= 60) {
+        value <- secDiff
+        unit <- 'second'
+    }
+
+    secBin <- ceiling(value)
+    binString <- paste0(secBin, unit)
+    binPer <- unitToPeriod(binString)
+    binPer
 }